WIP notes for current set management
